# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A service wrapping Interplanetary Filesystem which can be used to store and retrieve files there.",
        "title": "IPFS Document Manager API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
        "/create": {
            "post": {
                "description": "create document",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Create a new document in ipfs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of host tenant",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Document data to be created",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "list documents",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Get documents' identifiers from ipfs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of host tenant",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "get document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Get document from ipfs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of host tenant",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CID of stored document",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Delete a document from ipfs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of host tenant",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CID of document to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    }
                }
            }
        },
        "/{id}/update": {
            "put": {
                "description": "update document",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Update a document in ipfs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of host tenant",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CID of stored document",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Document data to be updated with",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.IpfsErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.IpfsErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.IpfsListResponse": {
            "type": "object",
            "properties": {
                "identifiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.DataIdentifier"
                    }
                }
            }
        },
        "main.IpfsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "identifier": {
                    "$ref": "#/definitions/types.DataIdentifier"
                }
            }
        },
        "types.DataIdentifier": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}